RegisterIndex now uses 64 bit for the address. Previous it used 8 bit, causing problems when running programs with ParameterType::Indirect.

+++ rust_project/loda-rust-core/src/execute/node_calc.rs
@@ -61,14 +61,14 @@ impl Node for NodeCalc {
         // TODO: deal with indirect
         match self.target.parameter_type {
             ParameterType::Register | ParameterType::Indirect => {
-                let value: u8 = u8::try_from(self.target.parameter_value).unwrap_or(255);
+                let value: u64 = u64::try_from(self.target.parameter_value).unwrap_or(255);
                 register_vec.push(RegisterIndex(value));
             },
             ParameterType::Constant => {}
         }
         match self.source.parameter_type {
             ParameterType::Register | ParameterType::Indirect => {
-                let value: u8 = u8::try_from(self.target.parameter_value).unwrap_or(255);
+                let value: u64 = u64::try_from(self.target.parameter_value).unwrap_or(255);
                 register_vec.push(RegisterIndex(value));
             },
             ParameterType::Constant => {}
